third_party/blink/web_tests/media/media-fragments/TC0005-expected.txt


Title: TC0005
Fragment: 't=3,7'
Comment: The media is requested from a to b.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
